modularization overhaul

+++ packages/knowhow/src/plugins/downloader/downloader.ts
@@ -2,14 +2,45 @@ import * as fs from "fs";
 import * as path from "path";
 import ytdl from "youtube-dl-exec";
 import Logger from "progress-estimator";
-import { DownloadInfo, KeyframeInfo } from "./types";
+import { DownloadInfo, KeyframeInfo, TranscriptChunk } from "./types";
 import { visionTool } from "../../agents/tools/visionTool";
 import { execAsync, fileExists, readFile, mkdir } from "../../utils";
-import { openai } from "../../ai";
+import { Clients } from "../../clients";
+import { Models } from "../../types";
 
 const logger = Logger();
 
-class DownloaderService {
+export class DownloaderService {
+  constructor(private clients: typeof Clients) {}
+
+  async askGptVision(
+    imageUrl: string,
+    question: string,
+    provider = "openai",
+    model = Models.openai.GPT_4o
+  ) {
+    const response = await this.clients.createCompletion(provider, {
+      model,
+      max_tokens: 2500,
+      messages: [
+        {
+          role: "user",
+          content: [
+            { type: "text", text: question },
+            {
+              type: "image_url",
+              image_url: {
+                url: imageUrl,
+              },
+            },
+          ],
+        },
+      ],
+    });
+
+    return response;
+  }
+
   async download(url: string, outputDir: string) {
     const info = await this.info(url);
     const exists = await fileExists(`${outputDir}/${info.id}.${info.ext}`);
@@ -48,17 +79,21 @@ class DownloaderService {
     // create a temp directory
     const outputDirPath = path.join(outputDir, `${fileName}/chunks`);
     await fs.promises.mkdir(outputDirPath, { recursive: true });
+    const doneFilePath = path.join(outputDirPath, ".chunking_done");
+
+    const doneFileExists = await fileExists(doneFilePath);
     const existingFolderFiles = await fs.promises.readdir(outputDirPath);
     const existingChunkNames = existingFolderFiles.filter(
       (f) => f.includes("chunk") && f.endsWith(".mp3")
     );
 
-    if (existingChunkNames.length > 0) {
+    if (existingChunkNames.length > 0 && doneFileExists) {
       if (reuseExistingChunks) {
         console.log("Chunks already exist, skipping");
-        return existingFolderFiles.map((chunkName) =>
+        const names = existingChunkNames.map((chunkName) =>
           path.join(outputDirPath, chunkName)
         );
+        return names;
       } else {
         for (const file of existingFolderFiles) {
           fs.rmSync(path.join(outputDirPath, file), { recursive: true });
@@ -66,8 +101,9 @@ class DownloaderService {
       }
     }
 
-    const command = `ffmpeg -i "${filePath}" -f segment -segment_time ${CHUNK_LENGTH_SECONDS} -map 0:a:0 -acodec mp3 -vn "${outputDirPath}/chunk%03d.mp3"`;
+    const command = `ffmpeg -i "${filePath}" -f segment -segment_time ${CHUNK_LENGTH_SECONDS} -map 0:a:0 -acodec mp3 -vn "${outputDirPath}/chunk%04d.mp3"`;
     await execAsync(command);
+    await fs.promises.writeFile(doneFilePath, "done");
 
     const folderFiles = await fs.promises.readdir(outputDirPath);
     const chunkNames = folderFiles.filter(
@@ -77,19 +113,23 @@ class DownloaderService {
     return chunkNames.map((chunkName) => path.join(outputDirPath, chunkName));
   }
 
-  public async transcribeChunks(
+  public async *streamTranscription(
     files: string[],
     outputPath: string,
     reusePreviousTranscript = true
-  ) {
+  ): AsyncGenerator<TranscriptChunk> {
     const exists = await fileExists(outputPath);
     if (exists && reusePreviousTranscript) {
-      console.log("Transcription already exists, skipping");
+      console.log("Transcription already exists, using cached data");
       const contents = await readFile(outputPath);
-      return JSON.parse(contents.toString()) as string[];
+      const data = JSON.parse(contents.toString()) as TranscriptChunk[];
+      for (const item of data) {
+        yield item;
+      }
+      return;
     }
 
-    const fullText = [];
+    const allTranscripts = [];
     for (const file of files) {
       const chunkName = path.parse(file).name;
       const chunkTranscriptPath = path.join(
@@ -99,16 +139,28 @@ class DownloaderService {
       const chunkExists = await fileExists(chunkTranscriptPath);
 
       if (chunkExists && reusePreviousTranscript) {
-        console.log("Chunk transcription already exists, skipping");
+        console.log(
+          chunkTranscriptPath,
+          " transcription already exists, using cached data"
+        );
         const contents = await readFile(chunkTranscriptPath);
-        fullText.push(contents.toString());
+        const cached = {
+          chunkPath: chunkTranscriptPath,
+          text: contents.toString(),
+          usd_cost: 0,
+        };
+
+        yield cached;
+        allTranscripts.push(cached);
         continue;
       }
 
       console.log("Transcribing", file);
-      const transcript = await openai.audio.transcriptions
-        .create({
-          file: fs.createReadStream(file),
+      const fileBuffer = fs.readFileSync(file);
+      const transcript = await this.clients
+        .createAudioTranscription("openai", {
+          file: fileBuffer,
+          fileName: path.basename(file),
           model: "whisper-1",
         })
         .catch((e) => {
@@ -120,36 +172,73 @@ class DownloaderService {
       await fs.promises.writeFile(chunkTranscriptPath, transcript.text);
 
       // save chunk transcript to file
-      fullText.push(transcript.text);
+      const data = {
+        chunkPath: chunkTranscriptPath,
+        text: transcript.text,
+        usd_cost: 30 * 0.0001, // assume 30 seconds,
+      };
+      yield data;
+      allTranscripts.push(data);
+    }
+
+    fs.writeFileSync(outputPath, JSON.stringify(allTranscripts, null, 2));
+  }
+
+  public async transcribeChunks(
+    files: string[],
+    outputPath: string,
+    reusePreviousTranscript = true
+  ): Promise<string[]> {
+    const exists = await fileExists(outputPath);
+    if (exists && reusePreviousTranscript) {
+      console.log("Transcription already exists, using cached data");
+      const contents = await readFile(outputPath);
+      return JSON.parse(contents.toString()) as string[];
+    }
+
+    const fullText = [];
+    for await (const { chunkPath, text } of this.streamTranscription(
+      files,
+      outputPath,
+      reusePreviousTranscript
+    )) {
+      console.log("Chunk transcribed:", chunkPath);
+      fullText.push(text);
     }
 
     await fs.promises.writeFile(outputPath, JSON.stringify(fullText));
     return fullText;
   }
 
-  public async extractKeyframes(
+  public async *streamKeyFrameExtraction(
     filePath: string,
-    outputPath: string,
+    videoJsonPath: string,
+    reusePreviousKeyframes: boolean = true,
     interval: number = 10
-  ): Promise<KeyframeInfo[]> {
-    if (fs.existsSync(outputPath)) {
-      console.log("Keyframes already exist, skipping");
-      const contents = await readFile(outputPath);
-      return JSON.parse(contents.toString()) as KeyframeInfo[];
+  ): AsyncGenerator<KeyframeInfo> {
+    if (reusePreviousKeyframes && fs.existsSync(videoJsonPath)) {
+      console.log("Keyframes already exist, using cached data");
+      const contents = await readFile(videoJsonPath);
+      const data = JSON.parse(contents.toString()) as KeyframeInfo[];
+      for (const keyframe of data) {
+        yield { ...keyframe, usd_cost: 0 };
+      }
+      return;
     }
 
     const parsed = path.parse(filePath);
-    const outputDir = path.dirname(outputPath);
+    const outputDir = path.dirname(videoJsonPath);
     const fileName = parsed.name;
     const keyframesDir = path.join(outputDir, `/keyframes`);
     await fs.promises.mkdir(keyframesDir, { recursive: true });
 
     const command = `ffmpeg -i "${filePath}" -vf "fps=1/${interval},scale=640:-1" "${keyframesDir}/frame%04d.jpg"`;
     await execAsync(command);
+    console.log("Extracting keyframe:", command);
 
     const keyframes = await fs.promises.readdir(keyframesDir);
-    const keyframeInfos: KeyframeInfo[] = [];
 
+    const allKeyframes = [];
     for (const keyframe of keyframes) {
       const keyframePath = path.join(keyframesDir, keyframe);
       const keyframeName = path.parse(keyframe).name;
@@ -159,46 +248,217 @@ class DownloaderService {
       );
       const descriptionExists = await fileExists(keyframeDescriptionPath);
 
-      if (descriptionExists) {
+      if (descriptionExists && reusePreviousKeyframes) {
         const cached = await readFile(keyframeDescriptionPath);
         const cachedJson = JSON.parse(cached.toString()) as KeyframeInfo;
-        keyframeInfos.push(cachedJson);
+        yield { ...cachedJson, usd_cost: 0 };
+        allKeyframes.push(cachedJson);
         continue;
       }
 
       const description = await this.describeKeyframe(keyframePath);
       const keyframeJson = {
         path: keyframePath,
-        description,
+        description: description.choices[0].message.content,
         timestamp: this.extractTimestamp(keyframe, interval),
+        usd_cost: description.usd_cost,
       };
       await fs.promises.writeFile(
         keyframeDescriptionPath,
         JSON.stringify(keyframeJson, null, 2)
       );
-      keyframeInfos.push(keyframeJson);
+      yield keyframeJson;
+      allKeyframes.push(keyframeJson);
     }
 
-    await fs.promises.writeFile(outputPath, JSON.stringify(keyframeInfos));
-
-    return keyframeInfos;
+    await fs.promises.writeFile(
+      videoJsonPath,
+      JSON.stringify(allKeyframes, null, 2)
+    );
   }
 
-  private async describeKeyframe(keyframePath: string): Promise<string> {
+  public async extractKeyframes(
+    filePath: string,
+    outputPath: string,
+    reusePreviousKeyframes: boolean = true,
+    interval: number = 10
+  ): Promise<KeyframeInfo[]> {
+    const keyframes: KeyframeInfo[] = [];
+    for await (const keyframe of this.streamKeyFrameExtraction(
+      filePath,
+      outputPath,
+      reusePreviousKeyframes,
+      interval
+    )) {
+      keyframes.push(keyframe);
+    }
+
+    await fs.promises.writeFile(outputPath, JSON.stringify(keyframes, null, 2));
+    return keyframes;
+  }
+
+  private async describeKeyframe(keyframePath: string) {
     const question =
       "Describe this image in detail, focusing on the main elements and actions visible.";
     const base64 = await fs.promises.readFile(keyframePath, {
       encoding: "base64",
     });
     const image = `data:image/jpeg;base64,${base64}`;
-    const description = await visionTool(image, question);
-    return description;
+    console.log("Describing keyframe:", keyframePath);
+    const response = await this.askGptVision(image, question);
+    return response;
   }
 
   private extractTimestamp(keyframeName: string, interval: number): number {
     const frameNumber = parseInt(keyframeName.match(/\d+/)[0], 10);
     return frameNumber * interval;
   }
-}
 
-export const Downloader = new DownloaderService();
+  async processAudio(
+    filePath: string,
+    reusePreviousTranscript = true,
+    chunkTime = 30
+  ): Promise<string[]> {
+    const parsed = path.parse(filePath);
+    const outputPath = `${parsed.dir}/${parsed.name}/transcript.json`;
+
+    // Skip chunking if the full output exists
+    const exists = await fileExists(outputPath);
+    if (exists && reusePreviousTranscript) {
+      console.log(
+        `Transcription ${outputPath} already exists, using cached data`
+      );
+      const fileContent = await readFile(outputPath, "utf8");
+      return outputPath.endsWith("txt")
+        ? fileContent.split("\n")
+        : JSON.parse(fileContent);
+    }
+
+    const chunks = await this.chunk(
+      filePath,
+      parsed.dir,
+      chunkTime,
+      reusePreviousTranscript
+    );
+    const transcription = await this.transcribeChunks(
+      chunks,
+      outputPath,
+      reusePreviousTranscript
+    );
+
+    return transcription;
+  }
+
+  async *streamProcessAudio(
+    filePath: string,
+    reusePreviousTranscript = true,
+    chunkTime = 30
+  ): AsyncGenerator<TranscriptChunk> {
+    const parsed = path.parse(filePath);
+    const outputPath = `${parsed.dir}/${parsed.name}/transcript.json`;
+
+    // Skip chunking if the full output exists
+    const exists = await fileExists(outputPath);
+    if (exists && reusePreviousTranscript) {
+      console.log(
+        `Transcription ${outputPath} already exists, using cached data`
+      );
+      const fileContent = await readFile(outputPath, "utf8");
+      const lines = outputPath.endsWith("txt")
+        ? fileContent.split("\n")
+        : JSON.parse(fileContent);
+
+      for (const line of lines) {
+        if (typeof line === "string") {
+          yield { chunkPath: "", text: line, usd_cost: 0 };
+        } else {
+          yield line as TranscriptChunk;
+        }
+      }
+      return;
+    }
+
+    const chunks = await this.chunk(
+      filePath,
+      parsed.dir,
+      chunkTime,
+      reusePreviousTranscript
+    );
+
+    for await (const chunk of this.streamTranscription(
+      chunks,
+      outputPath,
+      reusePreviousTranscript
+    )) {
+      yield chunk;
+    }
+  }
+
+  async processVideo(
+    filePath: string,
+    reusePreviousTranscript = true,
+    chunkTime = 30
+  ) {
+    const parsed = path.parse(filePath);
+    const outputPath = `${parsed.dir}/${parsed.name}/video.json`;
+
+    console.log("Processing audio...");
+    const transcriptions = await this.processAudio(
+      filePath,
+      reusePreviousTranscript,
+      chunkTime
+    );
+
+    console.log("Extracting keyframes...");
+    const videoAnalysis = await this.extractKeyframes(
+      filePath,
+      outputPath,
+      reusePreviousTranscript,
+      chunkTime
+    );
+
+    return videoAnalysis.map((frame, index) => {
+      return {
+        frame,
+        transcription: transcriptions[index],
+      };
+    });
+  }
+
+  async *streamProcessVideo(
+    filePath: string,
+    reusePreviousTranscript = true,
+    chunkTime = 30
+  ) {
+    const parsed = path.parse(filePath);
+    const videoJson = `${parsed.dir}/${parsed.name}/video.json`;
+
+    console.log("Processing audio...");
+    const transcriptions = this.streamProcessAudio(
+      filePath,
+      reusePreviousTranscript,
+      chunkTime
+    );
+
+    console.log("Extracting keyframes...");
+    const videoAnalysis = this.streamKeyFrameExtraction(
+      filePath,
+      videoJson,
+      reusePreviousTranscript,
+      chunkTime
+    );
+
+    for await (const frame of videoAnalysis) {
+      const transcription = (await transcriptions.next())
+        ?.value as TranscriptChunk;
+      yield {
+        frame,
+        transcription: transcription || {
+          chunkPath: "",
+          text: "[missing transcript]",
+          usd_cost: 0,
+        },
+      };
+    }
+  }
+}
